fastapi-service/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.28.0
pydantic==2.7.0
hypothesis==6.100.0
jsonschema==4.22.0
deepdiff==7.0.1
pyyaml==6.0.1
rich==13.7.0